pyproject.toml
[project]
name = "healthanalyticssystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "fastapi>=0.136.0",
    "joblib>=1.5.3",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.45.0",
    "xgboost>=3.2.0",
]
